moving the list display logic  out of HomeScreen -> listDisplay to potenemtially be reaused | overhaul serach to add our items and picture (still no good search)
