+++ app/screens/HomeScreen.jsx
@@ -10,7 +10,7 @@ import {
     TouchableOpacity,
     ActivityIndicator, RefreshControl,
 } from 'react-native';
-
+import {ListDisplay} from "@/components/listDisplay";
 import ItemCardTest from '@/components/itemCartTest';
 import { useFonts, FredokaOne_400Regular } from '@expo-google-fonts/fredoka-one';
 import { pullItem } from '@/app/services/item';
@@ -20,51 +20,6 @@ import {pullNonUserItem} from "@/app/services/nonUserService";
 
 export const HomeScreen = () => {
 
-
-    const [items, setItems] = useState([]);
-    const [loading, setLoading] = useState(true);
-    const [error, setError] = useState(null);
-    const [refreshing, setRefreshing] = useState(false);
-
-
-    const fetchData = async () => {
-        try {
-            const retrievedItems = await pullNonUserItem();
-            setItems(retrievedItems);
-            setLoading(false);
-        } catch (err) {
-            setError(err.message);
-            setLoading(false);
-        }
-    };
-
-
-    useEffect(() => {
-        fetchData();
-    }, []);
-
-    const onRefresh = useCallback(async () => {
-        setRefreshing(true);
-        await fetchData();
-        setRefreshing(false);
-    }, []);
-
-
-    if (loading) {
-        return (
-            <SafeAreaView style={styles.container}>
-                <ActivityIndicator size="large" color="#0000ff" />
-            </SafeAreaView>
-        );
-    }
-
-    if (error) {
-        return (
-            <SafeAreaView style={styles.container}>
-                <Text>Error: {error}</Text>
-            </SafeAreaView>
-        );
-    }
     return (
         <SafeAreaView style={styles.container}>
             <View style={{
@@ -82,27 +37,9 @@ export const HomeScreen = () => {
                 >What are you looking for today?</Text>
                 {/*<Text style={{color: '#545E66', fontSize: 24, transform: [{translateY: -50}], marginLeft: 15}}>Filter*/}
                 {/*    Placeholder</Text>*/}
+
             </View>
-            <FlatList
-                data={items}
-                keyExtractor={(item) => item.id.toString()}
-                renderItem={({item}) => <ItemCardTest item={item}/>}
-                refreshControl={
-                    <RefreshControl refreshing={refreshing} onRefresh={onRefresh} />
-                }
-                numColumns={2}
-                columnWrapperStyle={styles.columnWrapper}
-                contentContainerStyle={{
-                    paddingBottom: 50,
-                    flexGrow: 1,
-                    // shadowColor: "#000",
-                    // shadowOpacity: 0.2,
-                    // shadowRadius: 3,
-                    // elevation: 3,
-                    // shadowOffset: {width: -2, height: 4},
-                }}
-                showsVerticalScrollIndicator={false}
-            />
+            <ListDisplay />
         </SafeAreaView>
     );
 }
@@ -112,10 +49,10 @@ const styles = StyleSheet.create({
         flex: 1,
         backgroundColor: "#fff",
     },
-    columnWrapper: {
-        justifyContent: "space-between",
-        paddingHorizontal: 0,
-    },
+    // columnWrapper: {
+    //     justifyContent: "space-between",
+    //     paddingHorizontal: 0,
+    // },
 });
