+++ app/screens/Search.jsx
@@ -1,15 +1,55 @@
 import * as React from "react";
-import { useState } from "react";
-import { View, Text, TextInput, FlatList, StyleSheet } from "react-native";
+import {useEffect, useState} from "react";
+import {View, Text, TextInput, FlatList, StyleSheet, SafeAreaView, ActivityIndicator} from "react-native";
 import { useNavigation } from "@react-navigation/native";
 import { ITEMS } from "../../data-temp/items";
 import ItemCardTest from "../../components/itemCartTest.jsx";
+import {pullNonUserItem} from "@/app/services/nonUserService"
 
 export default function Search() {
-  const [query, setQuery] = useState("");
-  const navigation = useNavigation();
+    const [items, setItems] = useState([]);
+    const [loading, setLoading] = useState(true);
+    const [error, setError] = useState(null);
+    const [refreshing, setRefreshing] = useState(false);
+    const [query, setQuery] = useState("");
+    const navigation = useNavigation();
 
-  const filteredItems = ITEMS.filter((item) =>
+    const fetchData = async () => {
+        try {
+            const retrievedItems = await pullNonUserItem();
+            setItems(retrievedItems);
+            setLoading(false);
+        } catch (err) {
+            setError(err.message);
+            setLoading(false);
+        }
+    };
+
+
+    useEffect(() => {
+        fetchData();
+    }, []);
+
+    if (loading) {
+        return (
+            <SafeAreaView style={styles.container}>
+                <ActivityIndicator size="large" color="#0000ff" />
+            </SafeAreaView>
+        );
+    }
+
+    if (error) {
+        return (
+            <SafeAreaView style={styles.container}>
+                <Text>Error: {error}</Text>
+            </SafeAreaView>
+        );
+    }
+
+
+
+
+  const filteredItems = items.filter((item) =>
       item.title.toLowerCase().includes(query.toLowerCase())
   );
